requirements.txt
certifi==2018.11.29
chardet==3.0.4
cssselect2==0.2.1
Django==2.2.6
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-classy-tags==0.8.0
django-colorfield==0.1.15
django-i18n-model==0.0.9
django-i18nfield==1.5.0
django-import-export==1.2.0
django-js-asset==1.1.0
django-leaflet==0.24.0
django-phonenumber-field==2.2.0
filetype==1.0.5
gunicorn==19.9.0
idna==2.8
imgkit==1.0.2
lxml==4.3.4
MarkupSafe==1.1.1
phonenumbers==8.10.4
Pillow==6.2.0
psycopg2==2.7.3.2
pytz==2019.3
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tinycss2==1.0.2
tzlocal==1.5.1
urllib3==1.24.1
virtualenv==16.4.3
webencodings==0.5.1
cloudinary=1.18.2
